issue #3: file renaming
